+++ libnczarr/zutil.c
@@ -801,8 +801,8 @@ NCZ_comma_parse(const char* s, NClist* list)
 	endp = strchr(p,',');
 	if(endp == NULL) endp = p + strlen(p);
 	slen = (endp - p);
-	if((s = malloc(slen+1)) == NULL) {stat = NC_ENOMEM; goto done;}
-	memcpy(s,p,slen);
+	if((s = malloc((size_t)slen+1)) == NULL) {stat = NC_ENOMEM; goto done;}
+	memcpy(s,p,(size_t)slen);
 	s[slen] = '\0';
 	if(nclistmatch(list,s,0)) {
 	    nullfree(s); /* duplicate */
@@ -1013,9 +1013,9 @@ NCZ_fixed2char(const void* fixed, char** charp, size_t count, int maxstrlen)
 	if(p[0] == '\0') {
 	    sp = NULL;
 	} else {
-	    if((sp = (unsigned char*)malloc(maxstrlen+1))==NULL) /* ensure null terminated */
+	    if((sp = (unsigned char*)malloc((size_t)maxstrlen+1))==NULL) /* ensure null terminated */
 	        return NC_ENOMEM; 
-	    memcpy(sp,p,maxstrlen);
+	    memcpy(sp,p,(size_t)maxstrlen);
 	    sp[maxstrlen] = '\0';
 	}
 	charp[i] = (char*)sp;
