+++ ncgen/util.c
@@ -5,6 +5,7 @@
  *********************************************************************/
 
 #include "includes.h"
+#include <stddef.h>
 
 /* Track primitive symbol instances (initialized in ncgen.y) */
 Symbol* primsymbols[PRIMNO];
@@ -12,7 +13,7 @@ Symbol* primsymbols[PRIMNO];
 char*
 append(const char* s1, const char* s2)
 {
-    int len = (s1?strlen(s1):0)+(s2?strlen(s2):0);
+    size_t len = (s1?strlen(s1):0)+(s2?strlen(s2):0);
     char* result = (char*)ecalloc(len+1);
     result[0] = '\0';
     if(s1) strcat(result,s1);
@@ -371,7 +372,7 @@ collectpath(Symbol* grp, List* grpstack)
 char*
 prefixtostring(List* prefix, char* separator)
 {
-    int slen=0;
+    size_t slen=0;
     int plen;
     int i;
     char* result;
@@ -441,7 +442,7 @@ prefixdup(List* prefix)
     int i;
     if(prefix == NULL) return listnew();
     dupseq = listnew();
-    listsetalloc(dupseq,listlength(prefix));
+    listsetalloc(dupseq, (size_t)listlength(prefix));
     for(i=0;i<listlength(prefix);i++) listpush(dupseq,listget(prefix,i));
     return dupseq;
 }
@@ -488,7 +489,7 @@ pooldup(const char* s)
 char*
 poolcat(const char* s1, const char* s2)
 {
-    int len1, len2;
+    size_t len1, len2;
     char* cat;
     if(s1 == NULL && s2 == NULL) return NULL;
     len1 = (s1?strlen(s1):0);
